hospital_management/requirement.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
